breakout model functions
