+++ configs/rnn_training_config.yaml
@@ -1,3 +1,3 @@
 learning_rate_decay: 0.85
 learning_rate: 1
-number_of_epochs: 20
+number_of_epochs: 1
