+++ configs/run_model_pipeline.yaml
@@ -1,21 +1,8 @@
 stages:
 - name: train_rnn_model
   corpus_type: wikitext2
-  embedding_type: glove.42B.300d
+  embedding_type: glove.6B.50d
   training_config_file: rnn_training_config.yaml
   model_config_file: rnn_model_config.yaml
   lstm_configs:
     - default
-    - att-emb
-    - res-att-emb
-    - ff-emb
-    - res-ff-emb
-- name: rnn_training_comparison
-  lstm_configs:
-    - default
-    - att-emb
-    - res-att-emb
-    - ff-emb
-    - res-ff-emb
-
-
